requirements.txt
streamlit
sentence-transformers
faiss-cpu
numpy
langchain-community
pypdf
pandas
openpyxl
